ml_flow_vcs_api/requirements.txt
django
djangorestframework
pygments
pillow
pylint
django-filter
drf-extensions
django-rest-swagger
django-cors-headers
djoser
mysqlclient